[Refactor] dto 위치 수정 #110
- getUserProfileByNickname()에 필요한 Dto의 위치를 User로 옮겼습니다.
