+++ back/web-app/src/users/user/dto/get-user-profile-by-nickname-req.dto.ts
